+++ CHANGELOG.md
@@ -11,7 +11,26 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.0.0-beta.225 (2019-12-05)
+## v3.0.0-beta.227 (2019-12-06)
+
+#### :bug: Bug Fix
+
+* Fixed `bImage.onError` logic
+
+## v3.0.0-beta.226 (2019-12-06)
+
+#### :bug: Bug Fix
+
+* Fixed `bImage` cache
+* Fixed `bImage.onError` logic
+* Fixed keep-alive with `activatedProp`
+
+## v3.0.0-beta.225 (2019-12-06)
+
+#### :boom: Breaking Change
+
+* Renamed interfaces `Opts` -> `Options`
+* Renamed `iData` request error `error` -> `requestError`
 
 #### :rocket: New Feature
 
@@ -21,6 +40,12 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 * [Fixed `b-virtual-scroll` reloadLast method](https://github.com/V4Fire/Client/pull/144)
 * [Fixed `b-virtual-scroll` reload method](https://github.com/V4Fire/Client/pull/144)
+* [Fixed `v-image` onImageLoaded method](https://github.com/V4Fire/Client/pull/145)
+* Fixed race conditions within `iData`
+
+#### :house: Internal
+
+* Refactoring
 
 ## v3.0.0-beta.224 (2019-12-03)